+++ docs/changelog.md
@@ -7,6 +7,7 @@ Version 0.0.3
  * **Breaking change**: MML `r` rest commands now sends key-off events.
  * Added MML `w` wait command.
  * Quantized MML notes emit `rest_keyoff` bytecode instructions.
+ * Long MML `r` rests and `w` waits are optimised into loops, ie `r1 r1 r1 r1 r1 r1 r1 r1 r1 r1` will be converted to `r1 [r%216]4`.
  * Fixed play from line-start/cursor buttons not working when the cursor is at the start of a line or in a comment.
  * Highlight subroutine calls in MML note tracking.